refactor(commands): move flatten operation, add test for variant_command_generator, change the call to notifier in _generate
